Final version:
-- added small functionality solution
-- added wider functionality solution
-- added executables for both solutions
