--- Hmwk_#1_DictCounter/dummy.txt
@@ -1,28 +1,0 @@
-The Tragedy of Hamlet, Prince of Denmark
-
-ACT I
-SCENE I. Elsinore. A platform before the castle.
-
-    FRANCISCO at his post. Enter to him BERNARDO 
-
-BERNARDO
-
-    Who's there?
-
-FRANCISCO
-
-    Nay, answer me: stand, and unfold yourself.
-
-BERNARDO'S
-
-    Long live the king!
-
-FRANCISCO
-
-    Bernardo?
-
-BERNARDO
-
-    He.
-
-FRANCISCO
